Migrate Logs to emotion
